don't allow null in managed_by_auth

+++ jupyterhub/alembic/versions/4621fec11365_manage_roles.py
@@ -22,9 +22,23 @@ def upgrade():
     for table in ['group_role_map', 'roles', 'service_role_map', 'user_role_map']:
         if table not in tables:
             continue
-        op.add_column(table, sa.Column('managed_by_auth', sa.Boolean(), nullable=True))
+        # create column and assign existing rows with False
+        # since they are not managed
+        op.add_column(
+            table,
+            sa.Column(
+                'managed_by_auth',
+                sa.Boolean(),
+                server_default=sa.sql.False_(),
+                nullable=False,
+            ),
+        )
 
 
 def downgrade():
+    engine = op.get_bind().engine
+    tables = sa.inspect(engine).get_table_names()
     for table in ['group_role_map', 'roles', 'service_role_map', 'user_role_map']:
+        if table not in tables:
+            continue
         op.drop_column(table, 'managed_by_auth')
